// web/src/shared/api/generated/model/studentUnitResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { StudentLessonSummaryResult } from './studentLessonSummaryResult';

export interface StudentUnitResult {
  id: string;
  name: string;
  subjectId: string;
  subjectName: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  servableCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  masteredCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  seenCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  masteryPercent: number | string;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
     */
  bestExamScorePercent: number | string | null;
  lessons: StudentLessonSummaryResult[];
}